Make sample bookmarks API actually function
