+++ src/app/services/bookmarks.ts
@@ -1,8 +1,11 @@
 // TODO: move to slices when API exists
 
 import { useState } from "react";
-import { isaacApi } from "../state";
-import { ContentSummaryDTO } from "../../IsaacApiTypes";
+
+// TODO 
+// - add these endpoints in the backend (see bookmarksApi.ts)
+// - replace types with more appropriate ones (if necessary)
+// - modify /questions endpoint to return a bookmarked field for each question
 
 export const useBookmarks = () => {
     const [bookmarks, setBookmarks] = useState<string[]>([]);
@@ -24,34 +27,3 @@ export const useBookmarks = () => {
 
     return { isBookmarked, bookmarkItem };
 };
-
-// TODO 
-// - add these endpoints in the backend
-// - replace types with more appropriate ones (if necessary)
-// - modify /questions endpoint to return a bookmarked field for each question
-
-export const bookmarksApi = isaacApi.injectEndpoints({
-    endpoints: (build) => ({
-        getBookmarks: build.query<ContentSummaryDTO[], void>({
-            query: () => ({
-                url: "/users/current_user/bookmarks",
-                method: "GET"
-            }),
-            transformResponse: (response: ContentSummaryDTO[]) => response || []
-        }),
-        bookmarkItem: build.mutation<void, {id: string}>({
-            query: ({id}) => ({
-                url: "/users/current_user/bookmarks",
-                method: "POST",
-                body: {id}
-            }),
-        }),
-        deleteBookmark: build.mutation<void, {id: string}>({
-            query: ({id}) => ({
-                url: "/users/current_user/bookmarks",
-                method: "DELETE",
-                body: {id}
-            }),
-        }),
-    }),
-});
