+++ src/app/state/index.ts
@@ -36,6 +36,7 @@ export * from "./slices/api/authorisationsApi";
 export * from "./slices/api/accountApi";
 export * from "./slices/api/gameboardApi";
 export * from "./slices/api/assignmentsApi";
+export * from "./slices/api/bookmarksApi";
 export * from "./slices/api/contentApi";
 export * from "./slices/api/groupsApi";
 export * from "./slices/api/emailApi";
